Split GFAz implementation into its own header
