+++ src/index_registry.cpp
@@ -64,7 +64,7 @@
 
 #include "io/save_handle_graph.hpp"
 
-#include "algorithms/gfa_to_handle.hpp"
+#include "algorithms/gfaz_to_handle.hpp"
 #include "algorithms/prune.hpp"
 #include "algorithms/component.hpp"
 #include "algorithms/find_translation.hpp"
